:technologist: feat: update annotations and autoloading
- fix file name for ClientAdmin and ScopeAdmin
- update annotations for magic properties/methods
