+++ src/Entities/AccessTokenEntity.php
@@ -18,20 +18,17 @@ use League\OAuth2\Server\Entities\Traits\EntityTrait;
 use League\OAuth2\Server\Entities\Traits\TokenEntityTrait;
 use SilverStripe\ORM\DataObject;
 use SilverStripe\ORM\ManyManyList;
-use SilverStripe\ORM\SS_List;
 use SilverStripe\Security\Member;
 
 /**
- * @property string Code
- * @property string Expiry
- * @property bool Revoked
- * @property int ClientID
- * @property int MemberID
- * @property SS_List ScopeEntities
- *
+ * @property string $Code
+ * @property string $Expiry
+ * @property bool $Revoked
+ * @property int $ClientID
+ * @property int $MemberID
  * @method ClientEntity Client()
- * @method Member       Member()
- * @method ManyManyList ScopeEntities()
+ * @method Member Member()
+ * @method ManyManyList|ScopeEntity[] ScopeEntities()
  */
 class AccessTokenEntity extends DataObject implements AccessTokenEntityInterface
 {
